Add automatic import-map generation across graphql@14|15|16 (#5)
* Upgrade dependencies and transitive dependencies

* Add separate installs for graphql v14, v15, and v16

* Rewrite import map generator to automatically resolve conflicts

This enables the import map generator to automatically trace
graphql v14, v15, and v16 and finds the most specific common
import declaration between all three of them, if possible. As
a bonus, it now excludes exports that aren't available in all
three versions.

* Add peerDependencies entry for graphql to restrict versions

* Delete import-map-overrides.json

* Add warning and remove importMapOverrides from plugin

* Update README

* Fix index files with extension addition

When an extension is added, we must ensure that we add it
to the index module and not to a directory.